Fix P7/P8 review issues: security, dead code, dimensions, DRY
- Fix safeCompare length leak: pad buffers to equal length before timingSafeEqual
- Extract shared handleModeration() to deduplicate admin hide/unhide routes
- Add audit logging for moderation actions
- Wire WriterIdentityClient into writer + reader dashboards (was dead code)
- Add NEYNAR_API_KEY to .env.example
- Fix OG image dimensions 1200x800 → 1200x630 (Farcaster recommended)
- Export RESERVE_LABEL from constants to deduplicate inline IS_TESTNET checks
