+++ .env.example
@@ -53,6 +53,11 @@ DEPLOYER_PRIVATE_KEY=
 NEXT_PUBLIC_APP_URL=http://localhost:3000
 
 # -----------------------------------------------------------------------------
+# Farcaster Identity (Neynar API — optional, for writer profile display)
+# -----------------------------------------------------------------------------
+NEYNAR_API_KEY=
+
+# -----------------------------------------------------------------------------
 # Admin (Content Moderation)
 # -----------------------------------------------------------------------------
 ADMIN_API_KEY=
