+++ lib/contracts/constants.ts
@@ -35,6 +35,9 @@ export const PLOT_TOKEN = (IS_TESTNET
   ? "0x4200000000000000000000000000000000000006"
   : "0x0000000000000000000000000000000000000000") as `0x${string}`;
 
+/** Human-readable label for the reserve token */
+export const RESERVE_LABEL = IS_TESTNET ? "WETH" : "$PLOT";
+
 // ---------------------------------------------------------------------------
 // Mint Club V2
 // ---------------------------------------------------------------------------
