+++ src/app/dashboard/reader/page.tsx
@@ -5,6 +5,7 @@ import { useAccount } from "wagmi";
 import { useQuery } from "@tanstack/react-query";
 import { supabase, type Donation } from "../../../../lib/supabase";
 import { ReaderPortfolio } from "../../../components/ReaderPortfolio";
+import { WriterIdentityClient } from "../../../components/WriterIdentityClient";
 import { formatUnits } from "viem";
 import { ConnectWallet } from "../../../components/ConnectWallet";
 
@@ -76,6 +77,9 @@ export default function ReaderDashboard() {
       <h1 className="text-accent text-2xl font-bold tracking-tight">
         Reader Dashboard
       </h1>
+      <p className="text-muted mt-2 text-sm">
+        <WriterIdentityClient address={address!} />
+      </p>
 
       <ReaderPortfolio />
